Chore : 위치 리팩토링 #43
