+++ src/main/java/com/example/finalproject/apitest/dto/CompanyDataResponse.java
@@ -1,5 +1,5 @@
 package com.example.finalproject.apitest.dto;
-import com.example.finalproject.apitest.entity.*;
+import com.example.finalproject.apitest.entity.equity.*;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
