+++ src/main/java/com/example/finalproject/apitest/entity/equity/DartFundUsage.java
@@ -1,19 +1,18 @@
-// 1. 일반사항 Entity
-package com.example.finalproject.apitest.entity;
+package com.example.finalproject.apitest.entity.equity;
 
 import jakarta.persistence.*;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 import lombok.AllArgsConstructor;
 import java.time.LocalDateTime;
-import java.time.LocalDate;
 
+// 4. 자금사용목적 Entity
 @Entity
-@Table(name = "dart_general_info")
+@Table(name = "dart_fund_usage")
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
-public class DartGeneralInfo {
+public class DartFundUsage {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
@@ -31,32 +30,11 @@ public class DartGeneralInfo {
     @Column(name = "corp_name", length = 200)
     private String corpName;
 
-    @Column(name = "sbd")
-    private LocalDate sbd; // 청약기일
+    @Column(name = "se", length = 200)
+    private String se; // 구분
 
-    @Column(name = "pymd")
-    private LocalDate pymd; // 납입기일
-
-    @Column(name = "sband")
-    private LocalDate sband; // 청약공고일
-
-    @Column(name = "asand")
-    private LocalDate asand; // 배정공고일
-
-    @Column(name = "asstd")
-    private LocalDate asstd; // 배정기준일
-
-    @Column(name = "exstk", columnDefinition = "TEXT")
-    private String exstk; // 신주인수권 행사대상증권
-
-    @Column(name = "exprc")
-    private Long exprc; // 신주인수권 행사가격
-
-    @Column(name = "expd", columnDefinition = "TEXT")
-    private String expd; // 신주인수권 행사기간
-
-    @Column(name = "rpt_rcpn", length = 14)
-    private String rptRcpn; // 주요사항보고서 접수번호
+    @Column(name = "amt")
+    private Long amt; // 금액
 
     @Column(name = "created_at")
     private LocalDateTime createdAt;
