+++ src/main/java/com/example/finalproject/apitest/entity/equity/DartSellerInfo.java
@@ -1,19 +1,18 @@
-package com.example.finalproject.apitest.entity;
+package com.example.finalproject.apitest.entity.equity;
 
 import jakarta.persistence.*;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 import lombok.AllArgsConstructor;
 import java.time.LocalDateTime;
-import java.time.LocalDate;
 
-// 4. 자금사용목적 Entity
+// 5. 매출인에관한사항 Entity
 @Entity
-@Table(name = "dart_fund_usage")
+@Table(name = "dart_seller_info")
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
-public class DartFundUsage {
+public class DartSellerInfo {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
@@ -31,11 +30,20 @@ public class DartFundUsage {
     @Column(name = "corp_name", length = 200)
     private String corpName;
 
-    @Column(name = "se", length = 200)
-    private String se; // 구분
+    @Column(name = "hdr", length = 200)
+    private String hdr; // 보유자
 
-    @Column(name = "amt")
-    private Long amt; // 금액
+    @Column(name = "rl_cmp", length = 200)
+    private String rlCmp; // 회사와의관계
+
+    @Column(name = "bfsl_hdstk")
+    private Long bfslHdstk; // 매출전보유증권수
+
+    @Column(name = "slstk")
+    private Long slstk; // 매출증권수
+
+    @Column(name = "atsl_hdstk")
+    private Long atslHdstk; // 매출후보유증권수
 
     @Column(name = "created_at")
     private LocalDateTime createdAt;
