+++ src/main/java/com/example/finalproject/apitest/repository/DartGeneralInfoRepository.java
@@ -1,10 +1,9 @@
 // 1. DartGeneralInfoRepository
 package com.example.finalproject.apitest.repository;
 
-import com.example.finalproject.apitest.entity.DartGeneralInfo;
+import com.example.finalproject.apitest.entity.equity.DartGeneralInfo;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
-import org.springframework.data.repository.query.Param;
 import org.springframework.stereotype.Repository;
 import java.time.LocalDate;
 import java.util.List;
