+++ src/main/java/com/example/finalproject/apitest/repository/DartRedemptionRightsRepository.java
@@ -1,10 +1,9 @@
 // 6. DartRedemptionRightsRepository
 package com.example.finalproject.apitest.repository;
 
-import com.example.finalproject.apitest.entity.DartRedemptionRights;
+import com.example.finalproject.apitest.entity.equity.DartRedemptionRights;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
-import org.springframework.data.repository.query.Param;
 import org.springframework.stereotype.Repository;
 import java.util.List;
