+++ src/main/java/com/example/finalproject/apitest/repository/DartSecuritiesInfoRepository.java
@@ -1,13 +1,10 @@
 // 2. DartSecuritiesInfoRepository
 package com.example.finalproject.apitest.repository;
 
-import com.example.finalproject.apitest.entity.DartSecuritiesInfo;
+import com.example.finalproject.apitest.entity.equity.DartSecuritiesInfo;
 import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.data.jpa.repository.Query;
-import org.springframework.data.repository.query.Param;
 import org.springframework.stereotype.Repository;
 import java.util.List;
-import java.util.Optional;
 
 @Repository
 public interface DartSecuritiesInfoRepository extends JpaRepository<DartSecuritiesInfo, Long> {
