+++ src/main/java/com/example/finalproject/apitest/repository/DartSellerInfoRepository.java
@@ -1,10 +1,9 @@
 // 5. DartSellerInfoRepository
 package com.example.finalproject.apitest.repository;
 
-import com.example.finalproject.apitest.entity.DartSellerInfo;
+import com.example.finalproject.apitest.entity.equity.DartSellerInfo;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
-import org.springframework.data.repository.query.Param;
 import org.springframework.stereotype.Repository;
 import java.util.List;
