+++ src/main/java/com/example/finalproject/apitest/repository/DartUnderwriterInfoRepository.java
@@ -1,10 +1,9 @@
 // 3. DartUnderwriterInfoRepository
 package com.example.finalproject.apitest.repository;
 
-import com.example.finalproject.apitest.entity.DartUnderwriterInfo;
+import com.example.finalproject.apitest.entity.equity.DartUnderwriterInfo;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
-import org.springframework.data.repository.query.Param;
 import org.springframework.stereotype.Repository;
 import java.util.List;
